+++ backend/config/routes.rb
@@ -16,7 +16,6 @@ Rails.application.routes.draw do
       end
       resources :books, only: [:index, :create, :show] do
         get 'index_for_admin', on: :collection
-        patch 'delete_book', on: :member
         post 'fetch_book_info', on: :collection
         get 'fetch_categories', on: :collection
         get 'fetch_locations', on: :collection
