+++ frontend/src/lib/api/admin.js
@@ -11,3 +11,7 @@ export const fetchUsersAdmin = () => {
 export const searchBooks = (free_word,category_id) => {
   return client.get(`/admin/books/search`, { params:{q : free_word, category :category_id}});
 };
+
+export const deleteBook = (book_id) => {
+  return client.delete(`/admin/books/${book_id}`);
+};
